requirements.txt
Django>=5.1
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-extensions
psycopg2-binary